split side collapsible into parts ( step 1)
